created settings reader
SettingsReader.java looks for a file called settings.txt in the
project's root folder. Settings are stored as key-value pairs of
strings. In settings.txt, each line which contains exactly one = sign is
interpreted as a key=value pair; each line with exactly no = signs is
interpreted as a key, and the value "true" is assigned to it
automatically. For now, the only setting used is "theorem-path", the
path to the directory where the theorems are stored. Theorem.java has
been updated to use the settings file for this, instead of using a
hard-coded value. If settings.txt cannot be found, SettingsReader.java
should complain and provide a default value for each setting.