Lines, rays, and circles now use their operators
Each Drawable can now produce an expression to represent itself. The
line-on, ray-on, and circle-on operators have been consolidated into the
on operator, and the circle-center and ray-endpoint operators have been
renamed to center and endpoint, respectively. Each place where the
SketchCanvas creates a new line, ray, or circle has been changed. One
theorem which used the affected operators has been changed.