Slight improvement in allDependentsExist()
The search now returns early if any dependent fails to exist. The
double-recursion isn't quite as bad as it looks because ||
short-circuits: If a dependent doesn't exist, none of its dependents are
checked. The only better way I could think of to do this would involve
keeping a list of all the drawables we have checked, to avoid checking
any one twice.