Initialize LatexHandler when the program loads
This makes the program take longer to start up, but prevents the long
pause when the first statement is created. Users expect programs to take
time to start, but want them to be very responsive afterward.

This commit also changes (fixes?) the indentation in LatexHandler: the
whole class was somehow pushed one tab to the right.