Remove line, ray, and circle operators
Now that line, ray, and circle are types, these operators have no
purpose. (In fact, when the expression parser tries to parse "circle c",
it now reads it as a VariableExpression of type CIRCLE, not as an
OperatorExpression.)

Also, I missed a theorem in the last commit; its variables have types
now too.