You can only pair variables and expressions of the same type.
In pop-up menus for the statement list and SketchCanvas, the
button to pair with a theorem variable is only enabled if
(1) a statement or drawable is selected, (2) a pairing is selected
in the theorem panel, and (3) the type of the statement or drawable
matches the type of the variable in the pairing.

The Drawable interface has a new method getType() which returns the
drawable's type; each class implementing Drawable has a suitable
implementation of this method.